server/requirements/common.txt
Celery==5.3.1
crispy-bootstrap5==0.7
Django==4.2.3
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==8.0.0
django-constance[database]==2.9.1
django-crispy-forms==2.0
django-dbbackup==4.0.2
django-ipware==5.0.0
djangorestframework==3.14.0
drf-chunked-upload==0.5.1
humanize==4.7.0
paramiko==3.2.0
psycopg2-binary==2.9.6
pymemcache==4.0.0
requests==2.31.0
sentry-sdk==1.28.1
thumbhash-python==1.0.0